requirements.txt
# Core Voice Cloning Dependencies
TTS>=0.22.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.21.0,<5.0.0

# AI/LLM APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Numerical Computing (Compatible versions)
numpy>=1.24.0,<2.0.0
pandas>=1.5.0,<2.0.0
scipy>=1.11.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
openai-whisper>=20231117

# Web Framework & API
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0
python-multipart>=0.0.6

# Visualization & UI
matplotlib>=3.9.0
streamlit>=1.28.0
gradio>=4.8.0

# Database & Storage
supabase>=2.0.0
python-dotenv>=1.0.0

# ML Acceleration
accelerate>=0.25.0
datasets>=2.15.0

# Audio Quality Metrics
jiwer>=3.0.0
pesq>=0.0.4
pystoi>=0.3.3
